+++ activerecord/lib/active_record/associations/builder/has_and_belongs_to_many.rb
@@ -87,7 +87,7 @@ module ActiveRecord::Associations::Builder
       join_model.table_name_resolver = habtm
       join_model.class_resolver      = lhs_model
 
-      join_model.add_left_association :left_side, class: lhs_model
+      join_model.add_left_association :left_side, anonymous_class: lhs_model
       join_model.add_right_association association_name, belongs_to_options(options)
       join_model
     end
@@ -107,7 +107,7 @@ module ActiveRecord::Associations::Builder
 
     def middle_options(join_model)
       middle_options = {}
-      middle_options[:class] = join_model
+      middle_options[:anonymous_class] = join_model
       middle_options[:source] = join_model.left_reflection.name
       if options.key? :foreign_key
         middle_options[:foreign_key] = options[:foreign_key]